[trello-12]: Moved usperpanel services to separated dir
